Module MenhirLib

module Convert : sig ... end

This module defines two APIs for parsers, namely the traditional API and the revised API, and offers conversions between these APIs.

module Engine : sig ... end

module EngineTypes : sig ... end

This module defines several types and module types that are used in the specification of the module Engine.

module ErrorReports : sig ... end

This module offers functions that can help produce syntax error messages.

module IncrementalEngine : sig ... end

module InspectionTableFormat : sig ... end

module InspectionTableInterpreter : sig ... end

module LexerUtil : sig ... end

module LinearizedArray : sig ... end

module PackedIntArray : sig ... end

This module allows packing an array of (small) integers inside a string, using less than one word of memory per array element.

module Printers : sig ... end

module RowDisplacementDecode : sig ... end

Row displacement aims to compress a two-dimensional table where some values are considered insignificant.

module StaticVersion : sig ... end

module TableFormat : sig ... end

module TableInterpreter : sig ... end