Module Convert.Simplified

This submodule offers simplified versions of the above conversions. In this simplified API, concrete triples are used.

val traditional2revised : ('token, 'semantic_value) traditional -> ('token * Lexing.position * Lexing.position, 'semantic_value) revised

traditional2revised converts a traditional parser, produced by ocamlyacc or Menhir, into a revised parser.

val revised2traditional : ('token * Lexing.position * Lexing.position, 'semantic_value) revised -> ('token, 'semantic_value) traditional

revised2traditional converts a revised parser to a traditional parser.