Module MenhirLib.IncrementalEngine

type position = Lexing.position

module type INCREMENTAL_ENGINE = sig ... end

This signature describes the incremental LR engine. When the engine is used in this mode, the user controls the lexer, and the parser suspends itself when it needs to read a new token.

module type SYMBOLS = sig ... end

This signature is a fragment of the inspection API that is made available to the user when --inspection is used. This fragment contains type definitions for symbols.

module type INSPECTION = sig ... end

This signature describes the inspection API that is made available to the user when --inspection is used.

module type EVERYTHING = sig ... end

This signature combines the incremental API and the inspection API.